Add function and loop to break lines from file.
After some experimenting with recursion, I created
a function that can take a string and return a string
with line-breaks so that no line is longer than eighty
characters.

A simple loop under the function takes the filename,
finds the file, and uses the function to produce an
echo of the newly formatted file contents.