Added a simple test folder containing one test.
The test folder allows for repository of tests to be run
straight from terminal.

The shell test runs Ottanta on a test Markdown file and
compares this against a file containing the expected result.

At this commit, the test fails as Ottanta is not yet
working as expected (mostly due to the space at the beginning
of every line where it has been broken from the start
of the original.